sgm-75: refactor: update database schema and add tmdbId column
-   Renamed `MovieLocal` to package `movie`
-   Increased database version to 3.
-   Added migration from version 2 to 3.
-   Added `tmdb_id` column to the `movies` table.
- Updated `AppDatabase` to include the new migration.
