+++ data/src/main/java/io/kikiriki/sgmovie/data/AppDatabase.kt
@@ -2,11 +2,11 @@ package io.kikiriki.sgmovie.data
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import io.kikiriki.sgmovie.data.model.MovieLocal
+import io.kikiriki.sgmovie.data.model.movie.MovieLocal
 import io.kikiriki.sgmovie.data.repository.movie.local.MovieDao
 
 @Database(
-    version = 2,
+    version = 3,
     entities = [MovieLocal::class]
 )
 abstract class AppDatabase : RoomDatabase() {
