+++ data/src/main/java/io/kikiriki/sgmovie/data/repository/movie/local/MovieDao.kt
@@ -5,7 +5,7 @@ import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import androidx.room.Update
-import io.kikiriki.sgmovie.data.model.MovieLocal
+import io.kikiriki.sgmovie.data.model.movie.MovieLocal
 import kotlinx.coroutines.flow.Flow
 
 @Dao
